refactor : ReservationData 패키지 위치 변경, 네이밍 변경
