+++ src/test/java/com/prgrms/catchtable/reservation/domain/ReservationTimeTest.java
@@ -2,7 +2,7 @@ package com.prgrms.catchtable.reservation.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.prgrms.catchtable.common.data.reservation.ReservationData;
+import com.prgrms.catchtable.reservation.fixture.ReservationFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,7 +11,7 @@ class ReservationTimeTest {
     @Test
     @DisplayName("예약 선점 여부 변경에 성공한다")
     void reversePreOccupied() {
-        ReservationTime reservationTime = ReservationData.getReservationTimeNotPreOccupied();
+        ReservationTime reservationTime = ReservationFixture.getReservationTimeNotPreOccupied();
         reservationTime.reversePreOccupied();
 
         assertThat(reservationTime.isPreOccupied()).isTrue();
@@ -20,7 +20,7 @@ class ReservationTimeTest {
     @Test
     @DisplayName("예약 여부 변경에 성공한다.")
     void reverseOccupied() {
-        ReservationTime reservationTime = ReservationData.getReservationTimeNotPreOccupied();
+        ReservationTime reservationTime = ReservationFixture.getReservationTimeNotPreOccupied();
         reservationTime.reverseOccupied();
 
         assertThat(reservationTime.isOccupied()).isTrue();
