+++ src/test/java/com/prgrms/catchtable/reservation/fixture/ReservationFixture.java
@@ -1,4 +1,4 @@
-package com.prgrms.catchtable.common.data.reservation;
+package com.prgrms.catchtable.reservation.fixture;
 
 import static com.prgrms.catchtable.reservation.domain.ReservationStatus.COMPLETED;
 
@@ -10,7 +10,7 @@ import com.prgrms.catchtable.shop.domain.Shop;
 import java.time.LocalDateTime;
 import org.springframework.test.util.ReflectionTestUtils;
 
-public class ReservationData {
+public class ReservationFixture {
 
     public static Reservation getReservation(ReservationTime reservationTime) {
         return Reservation.builder()
